[FEAT] Schedules in event added, fixes in event crud

+++ controllers/event_controller.py
@@ -1,3 +1,4 @@
+from datetime import datetime, timedelta
 from fastapi.encoders import jsonable_encoder
 from config.db_connection import PostgresConnection
 from schemas.event_schema import Event, EventDisplay, EventEdit
@@ -23,7 +24,8 @@ class EventController():
                 start_time=row[5],
                 end_time=row[6],
                 tables=row[7],
-                img_url=row[8]
+                img_url=row[8],
+                meeting_duration=row[9]
             )
             result.append(event)
         return jsonable_encoder(result)
@@ -46,12 +48,13 @@ class EventController():
             start_time=row[5],
             end_time=row[6],
             tables=row[7],
-            img_url=row[8]
+            img_url=row[8],
+            meeting_duration=row[9]
         )
         return jsonable_encoder(event)
 
     def create(self, new_event: Event):
-        query = "INSERT INTO events (name, description, start_date, end_date, start_time, end_time, tables, img_url) VALUES (%s, %s, %s, %s, %s, %s, %s, %s) RETURNING event_id"
+        query = "INSERT INTO events (name, description, start_date, end_date, start_time, end_time, tables, img_url, meeting_duration) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s) RETURNING event_id"
         self.cursor.execute(query, (
             new_event.name,
             new_event.description,
@@ -60,11 +63,13 @@ class EventController():
             new_event.start_time,
             new_event.end_time,
             new_event.tables,
-            new_event.img_url
+            new_event.img_url,
+            new_event.meeting_duration
         ))
         last_id = self.cursor.fetchone()[0]
         row_affected = self.cursor.rowcount
         if not row_affected > 0:
+            self.cursor.close()
             return {
                 "error": True,
                 "details": "Event couldn't be created"
@@ -75,6 +80,27 @@ class EventController():
             table_list.append((table_num, last_id))
         query = "INSERT INTO tables_event (table_num, event_id) VALUES (%s, %s)"
         self.cursor.executemany(query, table_list)
+        row_affected = self.cursor.rowcount
+        if not row_affected > 0:
+            self.cursor.close()
+            return {
+                "error": True,
+                "details": "Tables couldn't be created"
+            }
+        # create schedule
+        schedule_list = []
+        start_date = datetime(new_event.start_date.year, new_event.start_date.month, new_event.start_date.day)
+        end_date = datetime(new_event.end_date.year, new_event.end_date.month, new_event.end_date.day)
+        while start_date <= end_date:
+            start_time_aux = datetime(start_date.year, start_date.month, start_date.day, new_event.start_time.hour, new_event.start_time.minute, new_event.start_time.second)
+            end_time_aux = datetime(start_date.year, start_date.month, start_date.day, new_event.end_time.hour, new_event.start_time.minute, new_event.end_time.second)
+            while start_time_aux < end_time_aux:
+                end_meeting = start_time_aux + timedelta(minutes=new_event.meeting_duration)
+                schedule_list.append((last_id, start_time_aux.time(), end_meeting.time(), start_time_aux.date()))
+                start_time_aux = end_meeting
+            start_date = start_date + timedelta(days=1)
+        query = "INSERT INTO schedules (event_id, start_time, end_time, day) VALUES (%s, %s, %s, %s)"
+        self.cursor.executemany(query, schedule_list)
         row_affected = self.cursor.rowcount
         self.cursor.close()
         if row_affected > 0:
@@ -101,15 +127,27 @@ class EventController():
                 "details": "Event not found"
             }
         # If the event exists delete all tables
-        query = "DELETE FROM tables_event WHERE event_id = %s"
-        self.cursor.execute(query, (event.event_id,))
-        row_affected = self.cursor.rowcount
-        if not row_affected > 0:
-            self.cursor.close()
-            return {
-                "error": True,
-                "details": "Tables couldn't be deleted"
-            }
+        if event.tables:
+            query = "DELETE FROM tables_event WHERE event_id = %s"
+            self.cursor.execute(query, (event.event_id,))
+            row_affected = self.cursor.rowcount
+            if not row_affected > 0:
+                self.cursor.close()
+                return {
+                    "error": True,
+                    "details": "Tables couldn't be deleted"
+                }
+        # If the event exists delete all schedules
+        if event.start_date and event.end_date and event.start_time and event.end_time and event.meeting_duration:
+            query = "DELETE FROM schedules WHERE event_id = %s"
+            self.cursor.execute(query, (event.event_id,))
+            row_affected = self.cursor.rowcount
+            if not row_affected > 0:
+                self.cursor.close()
+                return {
+                    "error": True,
+                    "details": "Schedules couldn't be deleted"
+                }
         # Update info
         query = "UPDATE events SET name = COALESCE(%s, name), description = COALESCE(%s, description), start_date = COALESCE(%s, start_date), end_date = COALESCE(%s, end_date), start_time = COALESCE(%s, start_time), end_time = COALESCE(%s, end_time), tables = COALESCE(%s, tables), img_url = COALESCE(%s, img_url) WHERE event_id = %s"
         self.cursor.execute(query, (
@@ -131,25 +169,47 @@ class EventController():
                 "details": "Event couldn't be updated"
             }
         # Create new tables
-        table_list = []
-        for table_num in range(1, event.tables + 1):
-            table_list.append((table_num, event.event_id))
-        query = "INSERT INTO tables_event (table_num, event_id) VALUES (%s, %s)"
-        self.cursor.executemany(query, table_list)
-        row_affected = self.cursor.rowcount
-        self.cursor.close()
-        if row_affected > 0:
-            self.conn.commit()
-            return {
-                "error": False,
-                "details": "Event updated successfully"
-            }
-        else:
-            return {
-                "error": True,
-                "details": "Event couldn't be updated"
-            }
-    
+        if event.tables:
+            table_list = []
+            for table_num in range(1, event.tables + 1):
+                table_list.append((table_num, event.event_id))
+            query = "INSERT INTO tables_event (table_num, event_id) VALUES (%s, %s)"
+            self.cursor.executemany(query, table_list)
+            row_affected = self.cursor.rowcount
+            if not row_affected > 0:
+                self.cursor.close()
+                return {
+                    "error": True,
+                    "details": "Tables couldn't be created"
+                }
+        # Create new schedule
+        if event.start_date and event.end_date and event.start_time and event.end_time and event.meeting_duration:
+            schedule_list = []
+            start_date = datetime(event.start_date.year, event.start_date.month, event.start_date.day)
+            end_date = datetime(event.end_date.year, event.end_date.month, event.end_date.day)
+            while start_date <= end_date:
+                start_time_aux = datetime(start_date.year, start_date.month, start_date.day, event.start_time.hour, event.start_time.minute, event.start_time.second)
+                end_time_aux = datetime(start_date.year, start_date.month, start_date.day, event.end_time.hour, event.start_time.minute, event.end_time.second)
+                while start_time_aux < end_time_aux:
+                    end_meeting = start_time_aux + timedelta(minutes=event.meeting_duration)
+                    schedule_list.append((event.event_id, start_time_aux.time(), end_meeting.time(), start_time_aux.date()))
+                    start_time_aux = end_meeting
+                start_date = start_date + timedelta(days=1)
+            query = "INSERT INTO schedules (event_id, start_time, end_time, day) VALUES (%s, %s, %s, %s)"
+            self.cursor.executemany(query, schedule_list)
+            row_affected = self.cursor.rowcount
+            self.cursor.close()
+            if not row_affected > 0:
+                return {
+                    "error": True,
+                    "details": "Event couldn't be updated"
+                }
+        self.conn.commit()
+        return {
+            "error": False,
+            "details": "Event updated successfully"
+        }
+
     def delete(self, event_id: int):
         # delete tables
         query = "DELETE FROM tables_event WHERE event_id = %s"
@@ -161,6 +221,10 @@ class EventController():
                 "error": True,
                 "details": "Tables couldn't be deleted"
             }
+        # delete user event participation
+        # TODO
+        # delete schedule
+        # TODO
         # delete event
         query = "DELETE FROM events WHERE event_id = %s"
         self.cursor.execute(query, (event_id,))
